+++ packages/viness-di/src/index.ts
@@ -1,6 +1,8 @@
-export * from './instantiation'
-export * from './instantiation-service'
-export * from './service-collection'
-export * from './service-registry'
-export * from './descriptors'
-export * from './container'
+import '@abraham/reflection';
+
+export * from './instantiation';
+export * from './instantiation-service';
+export * from './service-collection';
+export * from './service-registry';
+export * from './descriptors';
+export * from './container';