+++ packages/viness-di/src/util.ts
@@ -1,10 +1,12 @@
-import { ServiceIdentifier } from './instantiation'
+import { ServiceIdentifier } from './instantiation';
 
-export const serviceIds = new Map<string | Symbol, ServiceIdentifier<any>>()
+export const serviceIds = new Map<string | Symbol, ServiceIdentifier<any>>();
 
-export const DI_TARGET = '$di$target'
-export const DI_DEPENDENCIES = '$di$dependencies'
+export const DI_TARGET = '$di$target';
+export const DI_DEPENDENCIES = '$di$dependencies';
 
 export function getServiceDependencies(ctor: any): { id: ServiceIdentifier<any>; index: number; optional: boolean }[] {
-    return ctor[DI_DEPENDENCIES] || []
+    const deps = Reflect.getOwnMetadata(DI_DEPENDENCIES, ctor) || [];
+
+    return deps as any;
 }