+++ packages/viness-react/src/app/annotation/index.ts
@@ -1,1 +1,2 @@
-export * from './injectable'
+export * from './inject';
+export * from './injectable';