+++ packages/viness-react/src/app/annotation/inject.ts
@@ -1,5 +1,7 @@
 import { InjectableServiceId } from '../../token';
 
 export function Inject(id: InjectableServiceId): ParameterDecorator {
-    return (target: Object, propertyKey: string | symbol | undefined, parameterIndex: number) => {};
+    return (target: Object, propertyKey: string | symbol | undefined, parameterIndex: number) => {
+        console.log(target, propertyKey, parameterIndex);
+    };
 }