+++ packages/viness-react/src/store/action-dispatcher.ts
@@ -4,17 +4,15 @@ import { createToken } from '../token';
 import { IAction, IActionHandler, IActionMetadata } from './action.protocol';
 import { ACTION_METADATA } from './constants';
 
-export const IActionBus = createToken<ActionBus>('VinessActionBus');
+export const IActionDispatcher = createToken<ActionDispatcher>('VinessActionBus');
 
-@Injectable({ token: IActionBus })
-export class ActionBus {
+@Injectable({ token: IActionDispatcher })
+export class ActionDispatcher {
     dispatch<T extends IAction, R = any>(action: T): Promise<R> {
         const { constructor: actionType } = Object.getPrototypeOf(action);
 
         const { id } = Reflect.getMetadata(ACTION_METADATA, actionType) as IActionMetadata;
         const token = createToken(id);
-
-        console.log(id, token);
 
         const handler = ContaienrUtil.resolve(token) as IActionHandler<T>;