+++ packages/viness-react/src/store/hooks.ts
@@ -1,0 +1,4 @@
+import { useResolve } from '../hooks';
+import { ActionDispatcher, IActionDispatcher } from './action-dispatcher';
+
+export const useActionDispatcher = () => useResolve<ActionDispatcher>(IActionDispatcher);